backend/requirements.txt
Flask==3.1.3
flask-cors==6.0.2
gunicorn

flask
flask-cors
flask-sqlalchemy
flask-jwt-extended
werkzeug